requirements.txt
numpy
opencv-python
Pillow
pydub
python-dotenv
